add coffee cup
