--- lib/src/features/presenters/ui/atomic/card_first_top.dart
@@ -1,94 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:manga_easy_ranking/src/features/presenters/ui/atomic/card_image_perfil.dart';
-import 'package:manga_easy_ranking/src/features/presenters/ui/molecules/ranking_temporada_new.dart';
-import 'package:manga_easy_sdk/manga_easy_sdk.dart';
-import 'package:manga_easy_themes/manga_easy_themes.dart';
-
-class CardFirstTop extends StatelessWidget {
-  final NivelUser nivel;
-  const CardFirstTop({super.key, required this.nivel});
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Container(
-          margin: const EdgeInsets.symmetric(horizontal: 20),
-          decoration: const BoxDecoration(
-            borderRadius: BorderRadius.only(
-              topLeft: Radius.circular(12),
-              topRight: Radius.circular(12),
-            ),
-            color: mineShaft,
-          ),
-          padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 4),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text(
-                'Nivel ${nivel.lvl}',
-                style: Theme.of(context).textTheme.bodyLarge!.copyWith(
-                      fontWeight: FontWeight.bold,
-                      color: webOrange,
-                    ),
-              ),
-              Text(
-                '${nivel.quantity} XP',
-                style: Theme.of(context).textTheme.bodyLarge!.copyWith(
-                      fontWeight: FontWeight.bold,
-                      color: webOrange,
-                    ),
-              ),
-            ],
-          ),
-        ),
-        Container(
-          decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(18),
-            color: ThemeService.of.primaryColor,
-          ),
-          padding: const EdgeInsets.all(16),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              CircleAvatar(
-                radius: 28,
-                backgroundColor: webOrange,
-                child: CardImagePerfil(
-                  nivelUser: nivel,
-                  radius: 25,
-                ),
-              ),
-              Text(
-                // AppHelps.limitaTitle(nivel.name, 22),
-                nivel.name,
-                style: Theme.of(context).textTheme.titleLarge!.copyWith(
-                      fontWeight: FontWeight.bold,
-                      color: ThemeService.of.primaryText,
-                    ),
-              ),
-              Stack(
-                alignment: Alignment.center,
-                children: [
-                  // Image.asset(
-
-                    
-                  //   width: 35,
-                  //   height: 35,
-                  //   color: webOrange,
-                  // ),
-                  Text(
-                    '1',
-                    style: Theme.of(context).textTheme.titleLarge!.copyWith(
-                          fontWeight: FontWeight.bold,
-                          color: ThemeService.of.primaryColor,
-                        ),
-                  ),
-                ],
-              ),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-}
